+++ Dockerfile
@@ -1,7 +1,7 @@
 FROM python:3.12-slim
 
 COPY main.py pyproject.toml /
-COPY /bearification /bearification
+COPY /obearon /obearon
 RUN pip install .
 
 ENTRYPOINT ["python3", "-u", "main.py"]
